Create a Burst Decoder
Created functions and script to decode the burst events as found through find_deviations.py. The code automatically pulls in the burst event times that were previously calculated, and uses them as the bin starts and ends to calculate population firing rate vectors to send to the decoder, trained in the same way as bayes_replay_dependent_decoding.py where a GMM is trained on the binned individual epochs of the taste responses.

- burst_bayes_dependent_decoding.py: script to run the burst decoding
- functions/burst decoding_funcs.py: functions to decode the bursts by epoch as well as plot the bursts individually (same as in decoding_funcs.py just modified slightly).
- functions/decode_parallel.py: function added to take in individual burst firing rate vectors and decode.